+++ src/TimeSeries.jl
@@ -15,7 +15,7 @@ export TimeArray, AbstractTimeSeries,
        basecall,
        merge, collapse,
        readtimearray, writetimearray,
-       update, rename
+       update, rename, rename!
 
 @reexport using Dates