+++ src/timearray.jl
@@ -11,15 +11,15 @@ abstract type AbstractTimeSeries{T,N,D} end
 
 # Constructors
 
-    TimeArray(timestamp, values[, colnames, meta=nothing])
+    TimeArray(timestamp, values[, colnames, meta = nothing])
     TimeArray(ta::TimeArray; timestamp, values, colnames, meta)
-    TimeArray(data::NamedTuple, timestamp=:datetime, meta)
+    TimeArray(data::NamedTuple, timestamp = :datetime, meta)
 
 The second constructor will yields a new TimeArray with the new given fields.
 Note that the unchanged fields will be shared, there aren't any copy for the
 underlying arrays.
 
-The third constructor build a TimeArray from a NamedTuple.
+The third constructor builds a `TimeArray` from a `NamedTuple`.
 
 # Arguments
 
@@ -36,8 +36,11 @@ The third constructor build a TimeArray from a NamedTuple.
 
 # Examples
 
-    data = (datetime=[DateTime(2018, 11, 21, 12, 0), DateTime(2018, 11, 21, 13, 0)], col1=[10.2, 11.2], col2=[20.2, 21.2], col3=[30.2, 31.2])
-    ta = TimeArray(data; timestamp=:datetime, meta="Example")
+    data = (datetime = [DateTime(2018, 11, 21, 12, 0), DateTime(2018, 11, 21, 13, 0)],
+            col1 = [10.2, 11.2],
+            col2 = [20.2, 21.2],
+            col3 = [30.2, 31.2])
+    ta = TimeArray(data; timestamp = :datetime, meta = "Example")
 
 """
 struct TimeArray{T,N,D<:TimeType,A<:AbstractArray{T,N}} <: AbstractTimeSeries{T,N,D}
@@ -55,18 +58,19 @@ struct TimeArray{T,N,D<:TimeType,A<:AbstractArray{T,N}} <: AbstractTimeSeries{T,
             unchecked = false) where {T,N,D<:TimeType,A<:AbstractArray{T,N}}
         nrow = size(values, 1)
         ncol = size(values, 2)
+        colnames = copy(colnames)
 
-        unchecked && return new(timestamp, values, replace_dupes(colnames), meta)
+        unchecked && return new(timestamp, values, replace_dupes!(colnames), meta)
 
         nrow != length(timestamp) && throw(DimensionMismatch("values must match length of timestamp"))
         ncol != length(colnames) && throw(DimensionMismatch("column names must match width of array"))
 
         _issorted_and_unique(timestamp) && return new(
-            timestamp, values, replace_dupes(colnames), meta)
+            timestamp, values, replace_dupes!(colnames), meta)
 
         timestamp_r = reverse(timestamp)
         _issorted_and_unique(timestamp_r) && return new(
-            timestamp_r, reverse(values, dims = 1), replace_dupes(colnames), meta)
+            timestamp_r, reverse(values, dims = 1), replace_dupes!(colnames), meta)
 
         throw(ArgumentError("timestamps must be strictly monotonic"))
     end
@@ -89,7 +93,7 @@ TimeArray(ta::TimeArray;
           colnames = _colnames(ta), meta = _meta(ta), args...) =
     TimeArray(timestamp, values, colnames, meta; args...)
 
-function TimeArray(data::NamedTuple; timestamp::Symbol, meta=nothing, args...)
+function TimeArray(data::NamedTuple; timestamp::Symbol, meta = nothing, args...)
     columns = (key for key in keys(data) if key != timestamp)
     dat = hcat((data[key] for key in columns)...)
     TimeArray(data[timestamp], dat, collect(columns), meta; args...)
@@ -208,17 +212,20 @@ calculate the paging
     ret
 end
 
-function show(io::IO, ta::TimeArray{T}) where T
+function print_time_array(io::IO, ta::TimeArray{T}, short=false) where T
     # summary line
     nrow = size(values(ta), 1)
     ncol = size(values(ta), 2)
 
     print(io, "$(nrow)×$(ncol) $(typeof(ta))")
     if nrow != 0
-        println(io, " $(timestamp(ta)[1]) to $(timestamp(ta)[end])")
+        print(io, " $(timestamp(ta)[1]) to $(timestamp(ta)[end])")
     else  # e.g. TimeArray(Date[], [])
         return
     end
+
+    short && return
+    println(io)
 
     # calculate column withs
     drow, dcol = displaysize(io)
@@ -298,8 +305,17 @@ function show(io::IO, ta::TimeArray{T}) where T
         end
     end  # for p ∈ pages
 end
+Base.show(io::IO, ta::TimeArray) = print_time_array(io, ta, true)
+Base.show(io::IO, ::MIME"text/plain", ta::TimeArray) =
+    print_time_array(io, ta, false)
+
 
 ###### getindex #################
+
+# the getindex function should return a new TimeArray, and copy data from
+# the source, includes `timestamp`, `values` and `colnames`.
+
+getindex(ta::TimeArray) = throw(BoundsError(typeof(ta), []))
 
 # single row
 getindex(ta::TimeArray, n::Integer) =